src/day3/requirements.txt
pydantic>=2.9.0
anthropic>=0.40.0
psycopg2-binary>=2.9.0
llama-index-core>=0.12.0
llama-index-vector-stores-qdrant>=0.4.0
llama-index-embeddings-fastembed>=0.3.0
qdrant-client>=1.12.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
chainlit>=2.0.0
